requirements.txt
Flask>=2.0.0
python-dotenv
numpy
PyMuPDF
google-cloud-translate
deepl